give looper object responsibility for creating its own menu widgets

+++ include/midi/midi_mapper_matrix_types.h
@@ -1,11 +1,11 @@
 #ifndef MATRIX_MAPPER_TYPES__INCLUDED
 #define MATRIX_MAPPER_TYPES__INCLUDED
 
-typedef int source_id_t;
-typedef int target_id_t;
+typedef int8_t source_id_t;
+typedef int8_t target_id_t;
 
 class MIDIMatrixManager;
 
 extern MIDIMatrixManager *midi_matrix_manager;
 
-#endif
+#endif
